refactor: reuse topic cards
