+++ packages/shared/src/components/onboarding/index.ts
@@ -1,0 +1,2 @@
+export * from './FilterOnboarding';
+export * from './FilterOnboardingStep';
